Refactor code to update offset type in FeedItem struct

+++ src/feed/feed_item.rs
@@ -1,11 +1,11 @@
 use crate::media::core::Media;
 
 pub enum FeedItem {
-    Media { media: Media, feed_index: i32 },
+    Media { media: Media, feed_index: usize },
 }
 
-impl From<(Media, i32)> for FeedItem {
-    fn from(value: (Media, i32)) -> Self {
+impl From<(Media, usize)> for FeedItem {
+    fn from(value: (Media, usize)) -> Self {
         FeedItem::Media {
             media: value.0,
             feed_index: value.1,
